[Xposed] Invocation handler for hooked methods
Add logic to the proxy invocation handler so it can handler hooked methods.
Much of the code can be shared, it mainly differs in the Java handler method
calls and less strict checks (e.g. a hooked method can be static, a proxy
method cannot; exceptions are not wrapped).

The reference to the handler class/method is stored as a static reference
and is set in Xposed. This is not the best code style and could be exchanged
for getter/setter later (with no practical benefit though).

Change-Id: Idb46cba14c9d70d13c75bc0dcdcbbec3e7c72efc