[Xposed] Generate Xposed info for system apps
During Zygote startup, search through the system apps and framework
files. If they don't have Xposed info so far (either inside the .oat or
as external file), generate the external Xposed file as we don't except
to be able to write to /system and don't want to waste space by keeping
the compiled code twice.

Change-Id: Ie484631a39f5e6e2127b59417eba1b525d38a326