[Xposed] Samsung: Let Zygote create the Xposed info file
The Samsung stock kernel is compiled with additional patches that
prevent processes from running as root, they are executed as
uid 2000 (shell) instead. This user can't create files in the
dalvik-cache directory, therefore dex2oat --xposed-only fails
(for whetever reason only on arm64, but not on arm).

The solution is to let Zygote create the file. dex2oat will then try to
unlink it again (fails due to missing permissions, but is ignored) and
reopen the already created file. It will also fchown() the file,
therefore it's important that the file is owned by uid 2000.
Finally, it will use the file in the same way as usual.

Note that --oat-fd would usually be the better choice, but it's a single
parameter whereas the boot image needs multiple output files.

Change-Id: I6a90cd915e88a3707120f353ee49593a2ea223bc