[Xposed] Visit the "this" reference of proxy methods on the stack
This seems to be a bug in AOSP, therefore I have reported it to Google:
  https://issuetracker.google.com/issues/70216372

artQuickProxyInvokeHandler() wraps all raw pointers into local
references and fixes up the stack after the proxy method has returned.
However, while the proxy method is still running, the pointers on the
stack might become stale if these objects get moved to another address
by the Garbage Collector. For normal methods, the stack will be updated
when the GC visits all places where objects might be referenced, however
this isn't done for proxy methods as they don't have a stack map.

The consequence was that the stale pointer caused crashes when used,
like in the GC at a later time ("Tried to mark 0x... not contained by
any spaces").

This patch ensures that at least the "this" object, which is stored at a
well-known place, is visited. Therefore the reference on the stack will
be updated when moving objects. This doesn't cover other arguments to
the proxy methods, however they don't seem to be read off the stack
directly anywhere.

Change-Id: Ic3aff4bf454298336a8de471139b1983cae367b1